refactor(api/v2): extract system domain into its own package
Move the system-information domain handlers out of the monolithic facade
package api into a new internal/api/v2/system package, following the
established domain-extraction pattern (Handler embeds *apicore.Core by
pointer and exposes Register*Routes). This is the system slice of the
api/v2 split.

What moved into internal/api/v2/system (package system):
- System info endpoints (system.go): /system/info, /resources, /disks,
  /jobs, /processes, /temperature/cpu, /database/stats, /database/v2/stats,
  POST /database/backup, /network-interfaces, /restart-status, plus
  /system/models (system_models.go) and /system/inference
  (inference_status.go).
- Events (events.go, events_aggregation.go): /system/events/detections,
  /system/events/operational.
- Diagnostics (diagnostics.go): the health-check infrastructure and
  /system/diagnostics/* endpoints, with the HealthMetricsStore() /
  HealthEventBuffer() accessors.
- Metrics history (metrics_history.go): /system/metrics/history and the
  /system/metrics/stream SSE endpoint (rate limiter preserved verbatim).
- Browser terminal (terminal.go, terminal_unix.go, terminal_windows.go):
  /terminal/ws.

Handlers are receiver-retyped *Controller -> *Handler with bodies verbatim.
The four route initializers are kept as separate ordered Register*Routes
entries (system, terminal, diagnostics, metrics-history) at the same
ordinal positions in the facade routeInitializers slice, so route order is
preserved.

Inference / broadcaster: BroadcastInferenceTopologyChanged stays on the
facade (*Controller, system_routes.go) with its nil-receiver guard, because
promoting it to a *Core method would dereference the embedded core on a nil
*Controller before the guard could run; it is called externally by
internal/analysis. Only the GetInferenceStatus endpoint and its snapshot
helpers moved.

Health infrastructure: the five diagnostics health fields moved onto
system.Handler, which owns them. The facade keeps nil-guarded delegators
HealthMetricsStore() / HealthEventBuffer() so the analysis pipeline keeps
reading them through *apiv2.Controller. WithHealthErrorBuffer now seeds
c.healthErrorBuf, injected into system.New.

Lifecycle: the CPU sampler and metrics collector run on the shared core's
Go()/Context(), so the facade Shutdown() tears them down exactly as before.

Shared substrate promoted to apicore: the CPU cache (CPUCache,
UpdateCPUCache, GetCachedCPUUsage), FormatBytesUint64, MetricsCollectorInterval
and DBTypeLegacy/DBTypeV2; these are consumed by both the moved system
handlers and staying package-api code (the /health endpoint, backup.go,
database_overview.go).

Stays in package api: the /system/audio/* device handlers (audio_devices.go)
and external-media endpoint belong to the audio/streaming and media domains
and are registered by a trimmed initSystemRoutes (system_routes.go) until
those domains are extracted. process_cache.go stays as well: its
processingCache is consumed only by media.go and the cleanup lifecycle, not
by any system handler. apiv2.MetricsHistoryMaxPoints is re-exported as an
alias to keep internal/api/server.go compiling.

No public-behavior change: apiv2.Controller / New / NewWithOptions
signatures, the route set, all JSON contracts, per-route middleware, and the
external methods (BroadcastInferenceTopologyChanged, HealthMetricsStore,
HealthEventBuffer, MetricsHistoryMaxPoints) are unchanged. The
route-enumeration golden test is byte-identical.

Verification: go build ./... clean; go vet ./internal/api/v2/...
./internal/analysis/... clean; go test -race green for ./internal/api/v2/,
./internal/api/v2/system/, ./internal/api/v2/apicore/ and
./internal/analysis/...; golangci-lint 0 issues; route-enumeration golden
test identical.
