+++ internal/api/v2/system/terminal_test.go
@@ -1,4 +1,4 @@
-package api
+package system
 
 import (
 	"testing"
