+++ internal/api/v2/system/terminal_unix_test.go
@@ -1,6 +1,6 @@
 //go:build !windows
 
-package api
+package system
 
 import (
 	"testing"
